+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] - 21-10-2024
+
+### Added
+
+ - Upload and list artifacts that will be shared as dataset
+ - When verifying agreement additional check (policyEnforcement for agreement exists) added
+ 
+
 ## [0.1.1] - 09-10-2024
 
 ### Added